Update Command QMM012

+++ nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/command/wageprovision/statementitem/PaymentItemSetCommand.java
@@ -68,7 +68,7 @@ public class PaymentItemSetCommand {
 	/**
 	 * 限度金額
 	 */
-	private Integer limitAmount;
+	private Long limitAmount;
 
 	/**
 	 * 限度金額区分
